+++ resources/js/components/tables/InstantsearchProvider.vue
